add maven support; rename class hierarchy; add version metadata to jar; build cleanup
