+++ build.xml
@@ -20,7 +20,7 @@
   <target name="compile" depends="init" description="compile the source">
     <javac includeantruntime="false" srcdir="${src}" destdir="${build}" debug="true" debuglevel="lines,vars,source">
       <classpath>
-	<fileset dir="lib">
+        <fileset dir="lib">
           <include name="**/*.jar" />
         </fileset>
       </classpath>
@@ -35,16 +35,15 @@
     <manifestclasspath property="manifest_cp" jarfile="myjar.jar">
       <classpath refid="dep.runtime" />
     </manifestclasspath>
-    <jar jarfile="${dist}/Multiaddr.jar" basedir="${build}">
+    <jar jarfile="${dist}/multiaddr.jar" basedir="${build}">
       <manifest>
-	  <attribute name="Class-Path" value="${manifest_cp}"/>
+        <attribute name="Class-Path" value="${manifest_cp}"/>
+        <attribute name="Implementation-Vendor" value="io.ipfs"/>
+        <attribute name="Implementation-Title" value="multiaddr"/>
+        <attribute name="Implementation-Version" value="1.0.0"/>
       </manifest>
     </jar>
-    <copy todir=".">
-      <fileset file="${dist}/Multiaddr.jar"/>
-    </copy>
   </target>
-
 
   <target name="test" depends="compile,dist">
     <junit printsummary="yes" fork="true" haltonfailure="yes">
@@ -52,13 +51,14 @@
       <classpath>
 	<pathelement location="lib/junit-4.11.jar" />
 	<pathelement location="lib/hamcrest-core-1.3.jar" />
-	<pathelement location="Multiaddr.jar" />
+	<pathelement location="lib/multihash.jar" />
+	<pathelement location="dist/multiaddr.jar" />
       </classpath>
-      <test name="org.ipfs.api.MultiaddrTests" haltonfailure="yes">
+      <test name="io.ipfs.multiaddr.MultiAddressTest" haltonfailure="yes">
 	<formatter type="plain"/>
 	<formatter type="xml"/>
       </test>
-      </junit>
+    </junit>
   </target>
 
   <target name="clean" description="clean up">
