+++ src/main/java/io/ipfs/multiaddr/Base32.java
@@ -1,4 +1,4 @@
-package org.ipfs.api;
+package io.ipfs.multiaddr;
 
 import java.math.*;
