+++ src/main/java/io/ipfs/multiaddr/MultiAddress.java
@@ -1,4 +1,4 @@
-package org.ipfs.api;
+package io.ipfs.multiaddr;
 
 import java.io.*;
 import java.util.*;
